Use tuples instead of objects with a 'type' to identify oneOf
When 'oneOf' was used, it injected 'type' into the objects that it
could represent.

One problem was that 'type' could class with an existing object
property.

Another problem is that it assumed that everything in a 'oneOf' was an
object.

A third problem is that this doesn't work well if we want a binary
serialization.

By using a tuple instead, all these problems are fixed. At first, I
feared that writing schemas by hand wouldn't be as pleasant but it
turned out that it's actually nicer :-).

The generated typescript needed a newer version of typescript for the
narrowing to work properly.